Retrieve or register an R7 method for a generic

Generics partition a function into interface (a generic) and implementation (many methods). method<- allows you to register a method, an implementation for a specified class signature, with a generic.

method() retrieves a method for a given signature. You typically shouldn't need this function while programming, because calling the generic will automatically dispatch to the correct method, but it's often useful interactively in order to see the implementation of a specific method.

method(generic, signature)

method(generic, signature) <- value

Arguments

generic

A generic function.

signature

A method signature, a list of R7 class constructors (produced by new_class()) or names of S3 or S4 classes.

value

A function that implements the generic specification for the given signature. The arguments must be compatible with the generic.

Examples

# Create a generic
bizarro <- new_generic("bizarro", dispatch_args = "x")
# Register some methods
method(bizarro, "numeric") <- function(x, ...) rev(x)
method(bizarro, s3_class("factor")) <- function(x, ...) {
  levels(x) <- rev(levels(x))
  x
}
method(bizarro, s3_class("data.frame")) <- function(x, ...) {
  x[] <- lapply(x, bizarro)
  rev(x)
}

# Using a generic calls the methods automatically
bizarro(1)
#> [1] 1

# But it can be useful to explicitly retrieve a method in order to
# inspect its implementation
method(bizarro, "numeric")
#> <R7_method> method(bizarro, "numeric")
#> function (x, ...) 
#> rev(x)
#> <environment: 0x7ff242128dc8>
method(bizarro, s3_class("factor"))
#> <R7_method> method(bizarro, s3_class("factor"))
#> function (x, ...) 
#> {
#>     levels(x) <- rev(levels(x))
#>     x
#> }
#> <environment: 0x7ff242128dc8>