Register a R7 method for a generic

A generic defines the interface of a function. Once you have created a generic with new_generic(), you provide implementations for specific signatures by registering methods with method<-

The goal is for method<- to be the single function you need when working with R7 generics or R7 classes. This means that as well as registering methods for R7 classes on R7 generics, you can also register methods for R7 classes on S3 or S4 generics, and S3 or S4 classes for R7 generics. But this is not a general method registration function: at least one of generic and signature needs to be from R7.

method(generic, signature) <- value

Arguments

generic

A generic function, either created by new_generic(), new_external_generic(), or an existing S3 generic.

signature

A method signature. For R7 generics that use single dispatch, this should be one of the following:

An R7 class (created by new_class()).

An R7 union (created by new_union()).

An S3 class (created by s3_class()).

An S4 class (created by methods::getClass() or methods::new()).

A base type specified either with its constructor (logical, integer, double etc) or its name ("logical", "integer", "double" etc).

A base union type specified by its name: "numeric", "atomic", or "vector".

For R7 generics that use multiple dispatch, this can be a list of any of the above types.

For S3 generics, this must be an R7 class.

value

A function that implements the generic specification for the given signature.

Examples

# Create a generic
bizarro <- new_generic("bizarro", dispatch_args = "x")
# Register some methods
method(bizarro, "numeric") <- function(x) rev(x)
method(bizarro, s3_class("data.frame")) <- function(x) {
  x[] <- lapply(x, bizarro)
  rev(x)
}

# Using a generic calls the methods automatically
bizarro(head(mtcars))
#>                   carb gear am vs  qsec    wt drat  hp disp cyl  mpg
#> Mazda RX4            1    3  0  1 20.22 3.460 2.76 105  225   6 18.1
#> Mazda RX4 Wag        2    3  0  0 17.02 3.440 3.15 175  360   8 18.7
#> Datsun 710           1    3  0  1 19.44 3.215 3.08 110  258   6 21.4
#> Hornet 4 Drive       1    4  1  1 18.61 2.320 3.85  93  108   4 22.8
#> Hornet Sportabout    4    4  1  0 17.02 2.875 3.90 110  160   6 21.0
#> Valiant              4    4  1  0 16.46 2.620 3.90 110  160   6 21.0