Create a new R7 class

A class specifies the properties (data) that each of its objects will possess. The class, and its parent, determines which method will be used when an object is passed to a generic.

new_class(
  name,
  parent = R7_object,
  constructor = NULL,
  validator = function(x) NULL,
  properties = list()
)

Arguments

name
The name of the class, as a string.

parent
The parent class. Either a string or (better) an object constructor.

constructor
The constructor function. This is optional, unless you want to control which properties can be set on constructor.

validator
A function used to determine whether or not an object is valid. This is called automatically after construction, and whenever any property is set. It should return NULL if the object is valid, and otherwise return a character vector of problems.

properties
A list specifying the properties (data) that every object of the class will possess. Each property can either be a named string (specifying the class), or a call to new_property(), allowing greater flexibility.

Value

A object constructor, a function that can be used to create objects of the given class.

Examples

# Create an class that represents a range using a numeric start and end
range <- new_class("range",
  properties = list(
    start = "numeric",
    end = "numeric"
  )
)
r <- range(start = 10, end = 20)
r
#> <range> <R7_object>
#> @start 10
#> @end   20
# get and set properties with @
r@start
#> [1] 10
r@end <- 40
r@end
#> [1] 40

# Use a validator to ensure that start and end are both length 1,
# and that start is < end
range <- new_class("range",
  properties = list(
    start = "numeric",
    end = "numeric"
  ),
  validator = function(x) {
    if (length(x@start) != 1) {
      "@start must be a single number"
    } else if (length(x@end) != 1) {
      "@end must be a single number"
    } else if (x@end < x@start) {
      "@end must be great than or equal to @start"
    }
  }
)
try(range(start = c(10, 15), end = 20))
#> Error : Invalid <range> object:
#> - @start must be a single number
try(range(start = 20, end = 10))
#> Error : Invalid <range> object:
#> - @end must be great than or equal to @start
# Type validation is performed automatically in R7
try(range(start = "hello", end = 20))
#> Error : <range>@start must be of class <numeric>:
#> - `value` is of class <character>