Define a new generic

A generic function uses different implementations depending on the class of one or more arguments (the signature). Create a new generic with new_generic() then use method<- to add methods to it.

new_generic(name, fun = NULL, signature = NULL)

Arguments

name
The name of the generic. This should be the same as the object that you assign it to.

fun
An optional specification of the generic, which must call method_call() to dispatch to methods. This is usually generated automatically from the signature, but you may want to supply it if you want to add additional required arguments, or perform some standardised computation in the generic.

signature
A character vector providing the names of arguments to dispatch on. If omitted, defaults to the required arguments of fun.

See also

Examples

# A simple generic with methods for some base types and S3 classes
type_of <- new_generic("type_of", signature = "x")
method(type_of, "character") <- function(x, ...) "A character vector"
method(type_of, "data.frame") <- function(x, ...) "A data frame"
method(type_of, "function") <- function(x, ...) "A function"

type_of(mtcars)
#> [1] "A data frame"
type_of(letters)
#> [1] "A character vector"
type_of(mean)
#> [1] "A function"

# If you want to require methods implement additional arguments, supply
# them after ... in the call
mean2 <- new_generic("mean2", function(x, ..., na.rm = TRUE) {
   method_call()
})
method(mean2, "numeric") <- function(x, ..., na.rm = TRUE) {
  if (na.rm) {
    x <- x[!is.na(x)]
  }
  sum(x) / length(x)
}
method(mean2, "character") <- function(x, ...) {stop("Not supported")}