Get or set value of a property

prop() and @, gets the value of the given property, throwing an error if the property doesn't exist for that object.

prop_safely() returns NULL if a property doesn't exist, rather than throwing an error.

prop<- and @<- set a new value for the given property.

props() returns a list of all properties

props<- sets multiple properties at once, validating once after all are set.

prop_names() returns the names of the properties

prop_exists(x, "prop") returns TRUE iif x has property prop.

Usage

prop(object, name)

prop_safely(object, name)

prop_names(object)

props(object)

props(object) <- value

prop_exists(object, name)

prop(object, name, check = TRUE) <- value

object@name

Arguments

object

An object from a R7 class

name

The name of the parameter as a character. Partial matching is not performed.

value

For prop<-, a replacement value for the property; for props<-, a named list of values. The object is automatically checked for validity after the replacement is done.

check

If TRUE, check that value is of the correct type and run validate() on the object before returning.

Examples

horse <- new_class("horse", properties = list(
  name = "character",
  colour = "character",
  height = "numeric"
))
lexington <- horse(colour = "bay", height = 15, name = "Lex")
lexington@colour
#> [1] "bay"
prop(lexington, "colour")
#> [1] "bay"

lexington@height <- 14
prop(lexington, "height") <- 15

try(prop(lexington, "age"))
#> Error in prop(lexington, "age") : Can't find property <horse>@age
prop_safely(lexington, "age")
#> NULL