Get or set value of a property

property and @, gets the value of the given property, throwing an error if the property doesn't exist for that object.

property_safely returns NULL if a property doesn't exist, rather than throwing an error.

property<- and @<- set a new value for the given property.

property(object, name)

property_safely(object, name)

property(object, name, check = TRUE) <- value

object@name

Arguments

object
An object from a R7 class

name
The name of the parameter as a character. Partial matching is not performed.

check
If TRUE, check that value is of the correct type and run validate() on the object before returning.

value
A replacement value for the parameter. The object is automatically checked for validity after the replacement is done.

Examples

horse <- new_class("horse", properties = list(
  name = "character",
  colour = "character",
  height = "numeric"
))
lexington <- horse(colour = "bay", height = 15)
lexington@colour
#> [1] "bay"
property(lexington, "colour")
#> [1] "bay"

lexington@height <- 14
property(lexington, "height") <- 15

try(property(lexington, "age"))
#> Error : Can't find property <horse>@age
property_safely(lexington, "age")
#> NULL