Retrieve a method for an R7 generic

method() takes a generic and signature and retrieves the corresponding method. This is rarely needed because most of the time you'll rely on the the generic, via method_call(), to find and call the method for you. However, this introspection is useful if you want to see the implementation of a specific method.

method(generic, signature)

Arguments

generic

A generic function, either created by new_generic(), new_external_generic(), or an existing S3 generic.

signature

A method signature. For R7 generics that use single dispatch, this should be one of the following:

An R7 class (created by new_class()).

An R7 union (created by new_union()).

An S3 class (created by s3_class()).

An S4 class (created by methods::getClass() or methods::new()).

A base type specified either with its constructor (logical, integer, double etc) or its name ("logical", "integer", "double" etc).

A base union type specified by its name: "numeric", "atomic", or "vector".

For R7 generics that use multiple dispatch, this can be a list of any of the above types.

For S3 generics, this must be an R7 class.

Value

A function with class <R7_method>.

Examples

# Create a generic and register some methods
bizarro <- new_generic("bizarro", dispatch_args = "x")
method(bizarro, "numeric") <- function(x) rev(x)
method(bizarro, s3_class("factor")) <- function(x) {
  levels(x) <- rev(levels(x))
  x
}

# Printing the generic shows the registered method
bizarro
#> <R7_generic> function (x, ...)  with 3 methods:
#> 1: method(bizarro, "integer")
#> 2: method(bizarro, "double")
#> 3: method(bizarro, s3_class("factor"))

# And you can use method() to inspect specific implementations
method(bizarro, "integer")
#> <R7_method> method(bizarro, "integer")
#> function (x) 
#> rev(x)
#> <environment: 0x7fdd1573f1f0>
method(bizarro, s3_class("factor"))
#> <R7_method> method(bizarro, s3_class("factor"))
#> function (x) 
#> {
#>     levels(x) <- rev(levels(x))
#>     x
#> }
#> <environment: 0x7fdd1573f1f0>