Define a new generic

A generic function uses different implementations (methods) depending on the class of one or more arguments (the signature). Create a new generic with new_generic() then use method<- to add methods to it. The body of the generic always contains method_call(), which takes care of finding and calling the appropriate method.

new_generic(name, fun = NULL, dispatch_args = NULL)

method_call()

Arguments

name

The name of the generic. This should be the same as the object that you assign it to.

fun

An optional specification of the generic, which must call method_call() to dispatch to methods. This is usually generated automatically from the dispatch_args, but you may want to supply it if you want to add additional required arguments, or perform some standardised computation in the generic.

dispatch_args

A character vector giving the names of the arguments that form the signature, i.e. the arguments used for method dispatch.

If dispatch_args are omitted, but fun is supplied, will default to the arguments that appear before ... in fun. If there are no dots, it will default to the first argument. If both fun and dispatch_args are supplied, the dispatch_args must appear at the start of fun's formals.

Dispatch arguments

The arguments that are used to pick the method are called the dispatch arguments. In most cases, this will be one argument, in which case the generic is said to use single dispatch. If it consists of more than one argument, it's said to use multiple dispatch.

There are two restrictions on the dispatch arguments: they must be the first arguments to the generic and if the generic uses ..., it must occur immediately after the dispatch arguments.

See also

Examples

# A simple generic with methods for some base types and S3 classes
type_of <- new_generic("type_of", dispatch_args = "x")
method(type_of, "character") <- function(x, ...) "A character vector"
method(type_of, s3_class("data.frame")) <- function(x, ...) "A data frame"
method(type_of, "function") <- function(x, ...) "A function"

type_of(mtcars)
#> [1] "A data frame"
type_of(letters)
#> [1] "A character vector"
type_of(mean)
#> [1] "A function"

# If you want to require methods implement additional arguments, supply
# them after ... in the call
mean2 <- new_generic("mean2", function(x, ..., na.rm = TRUE) {
   method_call()
})
method(mean2, "numeric") <- function(x, ..., na.rm = TRUE) {
  if (na.rm) {
    x <- x[!is.na(x)]
  }
  sum(x) / length(x)
}
method(mean2, "character") <- function(x, ..., na.rm = TRUE) {
  stop("Not supported")
}