Define a new generic

Use new_generic to define a new generic for a package. If you need to reference a generic in another package either import the generic into your package's namespace or use new_external_generic to declare it as an external generic and add a method_register() call to your .onLoad function.

new_generic(name, signature = NULL, fun = NULL)

new_external_generic(package, generic, version = NULL)

Arguments

name
The name or generic object of the generic

signature
The signature of the generic

fun
The function to use as the body of the generic.

package
The package the external generic is defined in.

generic
The name or generic object of the generic

version
An optional version the package must meet for the method to be registered.