Define a new property

A property defines a named component of an object. Properties are typically used to store (meta) data about an object, and are often limited to a data of a specific class.

By specifying a getter and/or setter, you can make the property "dynamic" so that it's computed when accessed or has some non-standard behaviour when modified.

Usage

new_property(
  class = any_class,
  getter = NULL,
  setter = NULL,
  default = NULL,
  name = NULL
)

Arguments

class

If specified, any values must be one of these classes (or class union).

getter

An optional function used to get the value. The function should take self as its sole argument and return the value. If the property has a class the class of the value is validated.

If a property has a getter but doesn't have a setter, it is read only.

setter

An optional function used to set the value. The function should take self and value and return a modified object.

default

When an object is created and the property is not supplied, what should it default to? If NULL, defaults to the "empty" instance of class.

name

Property name, primarily used for error messages. Used primrarily for testing as it is set automatically when using a list of properties.

Examples

# Simple properties store data inside an object
pizza <- new_class("pizza", properties = list(
  slices = new_property("numeric", default = 10)
))
my_pizza <- pizza(slices = 6)
my_pizza@slices
#> [1] 6
my_pizza@slices <- 5
my_pizza@slices
#> [1] 5

your_pizza <- pizza()
your_pizza@slices
#> [1] 10

# Dynamic properties can compute on demand
clock <- new_class("clock", properties = list(
  now = new_property(getter = function(self) Sys.time())
))
my_clock <- clock()
my_clock@now; Sys.sleep(1)
#> [1] "2022-03-08 14:15:19 UTC"
my_clock@now
#> [1] "2022-03-08 14:15:20 UTC"
# This property is read only
try(my_clock@now <- 10)
#> Error : Can't set read-only property <clock>@now

# These can be useful if you want to deprecate a property
person <- new_class("person", properties = list(
  first_name = "character",
  firstName = new_property(
     getter = function(self) {
       warning("@firstName is deprecated; please use @first_name instead", call. = FALSE)
       self@first_name
     },
     setter = function(self, value) {
       warning("@firstName is deprecated; please use @first_name instead", call. = FALSE)
       self@first_name <- value
       self
     }
   )
))
hadley <- person(first_name = "Hadley")
hadley@firstName
#> Warning: @firstName is deprecated; please use @first_name instead
#> [1] "Hadley"
hadley@firstName <- "John"
#> Warning: @firstName is deprecated; please use @first_name instead
hadley@first_name
#> [1] "John"